Update README.md
This list is a comprehensive and curated collection of the best tools for bug bounty hunting and web application security testing. The purpose of this list is to provide a quick and practical reference for security researchers and penetration testers, helping them find the right tool for every stage of the penetration testing process, from reconnaissance and information gathering to vulnerability discovery and exploitation.